+++ examples/resolve_domain.rs
@@ -1,11 +1,15 @@
-use starknet::providers::SequencerGatewayProvider;
+use starknet::providers::{Provider, SequencerGatewayProvider};
 use starknet_core::types::FieldElement;
-use starknet_id::{address_to_domain, domain_to_address, ResolvingError, GOERLI_CONTRACT};
+use starknet_id::naming::{ResolvingError, GOERLI_CONTRACT, MAINNET_CONTRACT};
 
 #[tokio::main]
 async fn main() {
+    let client = SequencerGatewayProvider::starknet_alpha_mainnet();
     println!("On mainnet:");
-    let addr = domain_to_address!("th0rgal.stark").await;
+    let addr = client.domain_to_address(
+        "th0rgal.stark",
+        MAINNET_CONTRACT,
+    ).await;
     match addr {
         Ok(addr) => println!("address: {}", addr),
         Err(err) => match err {
@@ -15,11 +19,12 @@ async fn main() {
         },
     }
 
-    let domain_result = address_to_domain!(FieldElement::from_hex_be(
+    let domain_result = client.address_to_domain(
+        FieldElement::from_hex_be(
         "0x048F24D0D0618fa31813DB91a45d8be6c50749e5E19ec699092CE29aBe809294"
-    )
-    .unwrap())
-    .await;
+        ).unwrap(),
+        MAINNET_CONTRACT,
+    ).await;
     match domain_result {
         Ok(domain) => println!("domain: {}", domain),
         Err(err) => match err {
@@ -28,11 +33,11 @@ async fn main() {
             ResolvingError::InvalidDomain => println!("Invalid domain"),
         },
     }
+    let client_goerli = SequencerGatewayProvider::starknet_alpha_goerli();
 
     println!("On goerli:");
-    let addr = domain_to_address(
+    let addr = client_goerli.domain_to_address(
         "th0rgal.stark",
-        SequencerGatewayProvider::starknet_alpha_goerli(),
         GOERLI_CONTRACT,
     )
     .await;
@@ -45,17 +50,16 @@ async fn main() {
         },
     }
 
-    let domain_result = address_to_domain(
+    let domain_result = client.address_to_domain(
         FieldElement::from_hex_be(
             "0x048F24D0D0618fa31813DB91a45d8be6c50749e5E19ec699092CE29aBe809294",
         )
         .unwrap(),
-        SequencerGatewayProvider::starknet_alpha_goerli(),
         GOERLI_CONTRACT,
     )
     .await;
     match domain_result {
-        Ok(domain) => println!("domain: {}", domain),
+        Ok(domain_result) => println!("domain: {}", domain_result),
         Err(err) => match err {
             ResolvingError::ConnectionError(cause) => println!("Connection error: {}", cause),
             ResolvingError::InvalidContractResult => println!("Invalid contract result"),
