+++ starknet-providers/src/sequencer_gateway.rs
@@ -14,6 +14,11 @@ use starknet_core::{
         TransactionRequest, TransactionSimulationInfo, TransactionStatusInfo, TransactionTrace,
     },
 };
+use starknet_id::{
+    encoding::{decode, encode},
+    naming::{ResolvingError, SELECTOR_D2A, SELECTOR_A2D,},
+};
+
 use thiserror::Error;
 use url::Url;
 
@@ -630,7 +635,79 @@ impl Provider for SequencerGatewayProvider {
             }
         }
     }
+    
+    async fn domain_to_address(
+        &self,
+        domain: &str,
+        contract_addr: FieldElement,
+    ) -> Result<FieldElement, ResolvingError> {
+        if !domain.ends_with(".stark") {
+            return Err(ResolvingError::InvalidDomain);
+        }
+        let encoding_result = encode(&domain[0..domain.len() - 6]);
+        match encoding_result {
+            Ok(encoded) => {
+                match self
+                    .call_contract(
+                        CallFunction {
+                            contract_address: contract_addr,
+                            entry_point_selector: SELECTOR_D2A,
+                            calldata: vec![FieldElement::ONE, encoded],
+                        },
+                        BlockId::Latest,
+                    )
+                    .await
+                {
+                    Ok(result) => match result.result.first() {
+                        Some(x) => Ok(*x),
+                        None => Err(ResolvingError::InvalidContractResult),
+                    },
+                    Err(cause) => Err(ResolvingError::ConnectionError(cause.to_string())),
+                }
+            }
+            Err(_) => Err(ResolvingError::InvalidDomain),
+        }
+    }
+    
+    async fn address_to_domain(
+        &self,
+        address: FieldElement,
+        contract_addr: FieldElement,
+    ) -> Result<String, ResolvingError> {
+        match self
+            .call_contract(
+                CallFunction {
+                    contract_address: contract_addr,
+                    entry_point_selector: SELECTOR_A2D,
+                    calldata: vec![address],
+                },
+                BlockId::Latest,
+            )
+            .await
+        {
+            Ok(result) => {
+                let mut calldata = result.result.iter();
+                let mut domain = String::new().to_owned();
+                match calldata.next() {
+                    Some(_) => {
+                        calldata.for_each(|value| {
+                            domain.push_str(decode(*value).as_str());
+                            domain.push('.');
+                        });
+                        domain.push_str("stark");
+                        Ok(domain)
+                    }
+                    None => Err(ResolvingError::InvalidContractResult),
+                }
+            }
+            Err(cause) => Err(ResolvingError::ConnectionError(cause.to_string())),
+        }
+    }
+    
+    
 }
+
+
 
 // We need to manually implement this because `arbitrary_precision` doesn't work with `untagged`:
 //   https://github.com/serde-rs/serde/issues/1183
@@ -681,7 +758,7 @@ fn append_block_id(url: &mut Url, block_identifier: BlockId) {
 #[cfg(test)]
 mod tests {
     use starknet_core::types::StarknetErrorCode;
-
+    
     use super::*;
 
     #[test]
@@ -747,4 +824,6 @@ mod tests {
             _ => panic!("Unexpected result"),
         }
     }
+    
+    
 }
